feat: #1 pr 리뷰 반영 패키지 구조 수정 어노테이션 규칙 적용 필드명 안맞는것 수정
