+++ src/main/java/com/switcheese/server/comment/domain/model/Comment.java
@@ -1,0 +1,46 @@
+package com.switcheese.server.comment.domain.model;
+
+import com.switcheese.server.config.BaseEntity;
+import com.switcheese.server.keyboardSwitch.domain.model.KeyboardSwitch;
+import com.switcheese.server.member.domain.model.Member;
+import jakarta.persistence.Column;
+import jakarta.persistence.Entity;
+import jakarta.persistence.FetchType;
+import jakarta.persistence.GeneratedValue;
+import jakarta.persistence.GenerationType;
+import jakarta.persistence.Id;
+import jakarta.persistence.JoinColumn;
+import jakarta.persistence.ManyToOne;
+import jakarta.persistence.Table;
+import lombok.AccessLevel;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+
+@NoArgsConstructor(access = AccessLevel.PROTECTED)
+@Getter
+@Entity
+@Table(name = "comment")
+public class Comment extends BaseEntity {
+
+  @Id
+  @GeneratedValue(strategy = GenerationType.IDENTITY)
+  @Column(name = "comment_id")
+  private Long id;
+
+  @ManyToOne(fetch = FetchType.LAZY)
+  @JoinColumn(name = "member_id", nullable = false)
+  private Member member;
+
+  @ManyToOne(fetch = FetchType.LAZY)
+  @JoinColumn(name = "switch_id", nullable = false)
+  private KeyboardSwitch keyboardSwitch;
+
+  private String content;
+
+  private Integer likeCount;
+
+  private Integer dislikeCount;
+
+  private Boolean isDelete;
+
+}
