--- src/main/java/com/switcheese/server/domain/comment/CommentReaction.java
@@ -1,32 +1,0 @@
-package com.switcheese.server.domain.comment;
-
-import com.switcheese.server.domain.member.Member;
-import jakarta.persistence.*;
-import lombok.Data;
-import lombok.NoArgsConstructor;
-
-import java.time.LocalDateTime;
-@NoArgsConstructor
-@Data
-@Entity
-@Table(name = "comment_reaction")
-public class CommentReaction {
-
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long commentReactionId;
-
-    @ManyToOne(fetch = FetchType.LAZY)
-    @JoinColumn(name = "member_id", nullable = false)
-    private Member member;
-
-    @ManyToOne(fetch = FetchType.LAZY)
-    @JoinColumn(name = "comment_id", nullable = false)
-    private Comment comment;
-
-    private String likeType;
-
-    private Boolean isDelete;
-
-    private LocalDateTime createdAt;
-}
