--- src/main/java/com/switcheese/server/domain/member/MemberRepository.java
@@ -1,6 +1,0 @@
-package com.switcheese.server.domain.member;
-
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface MemberRepository extends JpaRepository<Member,Long>,MemberRepositoryCustom {
-}
