--- src/main/java/com/switcheese/server/domain/member/MemberRepositoryCustom.java
@@ -1,7 +1,0 @@
-package com.switcheese.server.domain.member;
-
-import java.util.List;
-
-public interface MemberRepositoryCustom {
-    public List<Member> test();
-}
