--- src/main/java/com/switcheese/server/domain/switcheese/Wish.java
@@ -1,32 +1,0 @@
-package com.switcheese.server.domain.switcheese;
-
-import com.switcheese.server.domain.member.Member;
-import com.switcheese.server.domain.switcheese.Switch;
-import jakarta.persistence.*;
-import lombok.Data;
-import lombok.NoArgsConstructor;
-
-import java.time.LocalDateTime;
-@NoArgsConstructor
-@Data
-@Entity
-@Table(name = "wish")
-public class Wish {
-
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long wishId;
-
-    @ManyToOne(fetch = FetchType.LAZY)
-    @JoinColumn(name = "member_id", nullable = false)
-    private Member member;
-
-    @ManyToOne(fetch = FetchType.LAZY)
-    @JoinColumn(name = "switch_id", nullable = false)
-    private Switch switchEntity;
-
-    private Boolean isDelete;
-
-    private LocalDateTime createdAt;
-}
-
