+++ src/main/java/com/switcheese/server/keyboardSwitch/domain/model/SwitchLike.java
@@ -1,0 +1,39 @@
+package com.switcheese.server.keyboardSwitch.domain.model;
+
+
+import com.switcheese.server.config.BaseEntity;
+import com.switcheese.server.member.domain.model.Member;
+import jakarta.persistence.Column;
+import jakarta.persistence.Entity;
+import jakarta.persistence.FetchType;
+import jakarta.persistence.GeneratedValue;
+import jakarta.persistence.GenerationType;
+import jakarta.persistence.Id;
+import jakarta.persistence.JoinColumn;
+import jakarta.persistence.ManyToOne;
+import jakarta.persistence.Table;
+import lombok.AccessLevel;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+
+@NoArgsConstructor(access = AccessLevel.PROTECTED)
+@Getter
+@Entity
+@Table(name = "switch_like")
+public class SwitchLike extends BaseEntity {
+
+  @Id
+  @GeneratedValue(strategy = GenerationType.IDENTITY)
+  @Column(name = "switch_like_id")
+  private Long switchLikeId;
+
+  @ManyToOne(fetch = FetchType.LAZY)
+  @JoinColumn(name = "member_id", nullable = false)
+  private Member member;
+
+  @ManyToOne(fetch = FetchType.LAZY)
+  @JoinColumn(name = "switch_id", nullable = false)
+  private KeyboardSwitch keyboardSwitchEntity;
+  private Boolean isDelete;
+
+}
