+++ src/main/java/com/switcheese/server/member/domain/MemberRepository.java
@@ -1,0 +1,8 @@
+package com.switcheese.server.member.domain;
+
+import com.switcheese.server.member.domain.model.Member;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface MemberRepository extends JpaRepository<Member, Long>, MemberRepositoryCustom {
+
+}
