+++ src/main/java/com/switcheese/server/member/domain/MemberRepositoryCustom.java
@@ -1,0 +1,10 @@
+package com.switcheese.server.member.domain;
+
+import com.switcheese.server.member.domain.model.Member;
+import java.util.List;
+
+public interface MemberRepositoryCustom {
+
+  public List<Member> test();
+
+}
